requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.4.0=pyhd8ed1ab_0
aiohttp=3.8.4=py39ha55989b_0
aiosignal=1.3.1=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
blas=2.115=mkl
blas-devel=3.9.0=15_win64_mkl
blinker=1.5=pyhd8ed1ab_0
brotli=1.0.9=h8ffe710_7
brotli-bin=1.0.9=h8ffe710_7
brotlipy=0.7.0=py39hb82d6ee_1004
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.12.7=h5b45459_0
cachetools=5.3.0=pyhd8ed1ab_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.0=py39h0878f49_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.1.3=win_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
cryptography=37.0.2=py39h7bc7c5c_0
cuda-cccl=12.1.55=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cudatoolkit=11.3.1=h59b6b97_2
cudnn=8.4.1.50=hf5f08ae_0
cycler=0.11.0=pyhd8ed1ab_0
fonttools=4.33.3=py39hb82d6ee_0
freetype=2.10.4=h546665d_1
frozenlist=1.3.3=py39ha55989b_0
gettext=0.19.8.1=ha2e2712_1008
glib=2.70.2=h7755175_4
glib-tools=2.70.2=h7755175_4
google-auth=2.16.2=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.42.0=py39hc60d5dd_0
gst-plugins-base=1.20.3=he07aa86_0
gstreamer=1.20.3=hdff456e_0
icu=70.1=h0e60522_0
idna=3.3=pyhd8ed1ab_0
imageio=2.25.0=pyh24c5eb1_0
importlib-metadata=6.0.0=pyha770c72_0
intel-openmp=2022.1.0=h57928b3_3787
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h8ffe710_1
kiwisolver=1.4.3=py39h2e07f2f_0
krb5=1.19.3=h1176d77_0
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libblas=3.9.0=15_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libcblas=3.9.0=15_win64_mkl
libclang=14.0.5=default_h77d9078_0
libclang13=14.0.5=default_h77d9078_0
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.2.56=0
libcurand-dev=10.3.2.56=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.12=h8ffe710_0
libffi=3.4.2=h8ffe710_5
libglib=2.70.2=h3be07f2_4
libiconv=1.16=he774522_0
liblapack=3.9.0=15_win64_mkl
liblapacke=3.9.0=15_win64_mkl
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libogg=1.3.4=h8ffe710_1
libpng=1.6.37=h1d00b33_2
libprotobuf=3.21.12=h12be248_0
libtiff=4.4.0=h2ed3b44_1
libuv=1.43.0=h8ffe710_0
libvorbis=1.3.7=h0e60522_0
libwebp=1.2.2=h57928b3_0
libwebp-base=1.2.2=h8ffe710_1
libxcb=1.13=hcd874cb_1004
libzlib=1.2.13=hcfcfb64_4
libzlib-wapi=1.2.13=hcfcfb64_4
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.2=py39ha55989b_0
matplotlib=3.5.2=py39hcbf5309_0
matplotlib-base=3.5.2=py39h581301d_0
mkl=2022.1.0=h6a75c08_874
mkl-devel=2022.1.0=h57928b3_875
mkl-include=2022.1.0=h6a75c08_874
msys2-conda-epoch=20160418=1
multidict=6.0.4=py39ha55989b_0
munkres=1.1.4=pyh9f0ad1d_0
numpy=1.23.0=py39h1a62c8c_0
oauthlib=3.2.2=pyhd8ed1ab_0
openjpeg=2.4.0=hb211442_1
openssl=1.1.1t=hcfcfb64_0
packaging=21.3=pyhd8ed1ab_0
pcre=8.45=h0e60522_0
pillow=9.1.1=py39ha53f419_1
pip=22.1.2=pyhd8ed1ab_0
protobuf=4.21.12=py39h99910a6_0
pthread-stubs=0.4=hcd874cb_1001
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.4=py39hb08f45d_1
pyqt5-sip=12.9.0=py39h415ef7b_1
pysocks=1.7.1=py39hcbf5309_5
python=3.9.13=h9a09f29_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.13.1=py3.9_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h16d0643_3
pytorch-mutex=1.0=cuda
pyu2f=0.1.5=pyhd8ed1ab_0
qt-main=5.15.4=h467ea89_2
requests=2.28.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.1.1=py39he931e04_0
scipy=1.8.1=py39hc0c34ad_0
setuptools=62.6.0=py39hcbf5309_0
sip=6.5.1=py39h415ef7b_2
six=1.16.0=pyh6c4a22f_0
sqlite=3.38.5=h8ffe710_0
tbb=2021.5.0=h2d74725_1
tensorboard=2.10.0=py39haa95532_0
tensorboard-data-server=0.6.1=py39haa95532_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h8ffe710_0
toml=0.10.2=pyhd8ed1ab_0
torchaudio=0.13.1=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.1=py39hb82d6ee_3
tqdm=4.65.0=pypi_0
typing-extensions=4.2.0=hd8ed1ab_1
typing_extensions=4.2.0=pyha770c72_1
tzdata=2022a=h191b570_0
ucrt=10.0.20348.0=h57928b3_0
unicodedata2=14.0.0=py39hb82d6ee_1
urllib3=1.26.9=pyhd8ed1ab_0
vc=14.2=hb210afc_6
vs2015_runtime=14.34.31931=h4c5c07a_10
werkzeug=2.2.3=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
win_inet_pton=1.1.0=py39hcbf5309_4
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.5=h62dcd97_1
yarl=1.8.2=py39ha55989b_0
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h6255e5f_1
